9/23打ち合わせをもとに、HxActions仕様変更。

+++ java/dao/src/jp/wda/reprorper/core/entity/Section.java
@@ -41,7 +41,7 @@ public class Section implements Serializable {
 	 */
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	public Long sectionId;
+	public int sectionId;
 
 	/**
 	 *
